test : NameTest 분리
